requirements.txt
pandas
requests
beautifulsoup4
googlesearch-python
openpyxl
